requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio

# Web Framework
fastapi>=0.136.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
starlette-sessions>=0.3.0

# AI & Orchestration (Notice: no 'langchain' base package)
langgraph>=1.1.0
langchain-core>=1.2.0
langchain-community>=0.4.0
langchain-openai>=1.1.0
langchain-pinecone>=0.2.0
langchain-huggingface>=1.2.0
langchain-classic==1.0.1
langchain-text-splitters>=1.1.0
sentence-transformers>=3.0.0
openai>=2.30.0

# Database & Infrastructure
psycopg2-binary>=2.9.0
pinecone>=7.0.0
boto3>=1.42.0  # This will bring in the correct botocore automatically
python-dotenv>=1.0.0
pypdf>=6.0.0
requests-oauthlib==2.0.0
itsdangerous>=2.2.0